log4j2; department view

+++ db/src/main/java/io/how2coder/demo/db/dao/DepartmentDao.java
@@ -2,5 +2,5 @@ package io.how2coder.demo.db.dao;
 
 import io.how2coder.demo.model.Department;
 
-public interface DepartmentDao extends ItemDao<Department, Integer> {
+public interface DepartmentDao extends ItemDao<Department, Long> {
 }
