+++ db/src/main/java/io/how2coder/demo/db/dao/ManagerDao.java
@@ -2,5 +2,5 @@ package io.how2coder.demo.db.dao;
 
 import io.how2coder.demo.model.Manager;
 
-public interface ManagerDao extends ItemDao<Manager,Integer> {
+public interface ManagerDao extends ItemDao<Manager, Long> {
 }
